Fix MacOS cleanup
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2700 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/classes/MethodClass.hpp
@@ -77,7 +77,7 @@ typedef uint16_t *(RexxEntry *PNATIVEMETHOD)(RexxMethodContext *, ValueDescripto
                                        /* pointer to native function function*/
 typedef uint16_t *(RexxEntry *PNATIVEROUTINE)(RexxCallContext *, ValueDescriptor *);
 
-typedef unsigned int (RexxEntry *PREGISTEREDROUTINE)(const char *, size_t, PCONSTRXSTRING, const char *, PRXSTRING);
+typedef size_t (RexxEntry *PREGISTEREDROUTINE)(const char *, size_t, PCONSTRXSTRING, const char *, PRXSTRING);
 
 class BaseExecutable : public RexxObject
 {
